requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.22
openai>=1.32.0
gunicorn==21.2.0
requests==2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
scipy>=1.11.0
